lxd/network/ovn: Add a wrapper for the ovsdb `(client).Transact` method
This commit adds logic to the client to listen for update events on the
`NB_Global` table. This is used to receive notifications when the value of
`sb_cfg` or `hv_cfg` is changed, so that we can wait for configuration to
propagate without polling.

The `(NB).transaction` method gets the existing NB_Global from the cache
and increments the `nb_cfg` value. It then waits for `sb_cfg` or `hv_cfg`
be updated to a number greater than or equal to the incremented `nb_cfg`.
This logic is the same as in ovn-nbctl (I've linked to the source in a
comment).

A new type `nbWaitMode` is added to control the wait behaviour.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit d66cee6c23c62ee4b197fe875e102ac12fce66ae)
Signed-off-by: Nikita Mezhenskyi <nikita.mezhenskyi@canonical.com>