test: Explicitly add network.ovs.connection to init_dump
This is needed because `spawn_lxd()` now sets `network.ovs.connection`
key for LXD. As a result, `init_dump` will contain this field as well,
even though the preseed file does not set it.
To make it more explicit, we include `network.ovs.connection` to the
preseed and the `expected.yaml`.

Signed-off-by: Nikita Mezhenskyi <nikita.mezhenskyi@canonical.com>